add blog-service.ts add post-service.ts
add try catch

+++ src/repositories/post-repository.ts
@@ -1,58 +1,74 @@
-
 import {OutputPostType, PostType} from "../types/posts/output";
-import {PostCreateModel, PostParams, PostUpdateModel} from "../types/posts/input";
-import {BlogRepository} from "./blog-repository";
-import {OutputBlogType} from "../types/blogs/output";
+import {PostUpdateModel} from "../types/posts/input";
 import {postCollection} from "../db/db";
 import {PostMapper} from "../types/posts/PostMapper";
 import {ObjectId, WithId} from "mongodb";
+import {isValidObjectId} from "./utils/Objcet(Id)Chek";
 
 export class PostRepository {
-    static async getAllPosts() : Promise<OutputPostType[]> {
-        const posts: WithId<PostType>[] = await  postCollection.find({}).toArray()
+    //Возвращает посты переработанные в мапере
+    static async getAllPosts(): Promise<OutputPostType[]> {
+        const posts: WithId<PostType>[] = await postCollection.find({}).toArray();
         return posts.map(PostMapper)
     }
 
-    static async getPostById(id: string) : Promise<OutputPostType | null> {
-        const post : WithId<PostType> | null = await postCollection.findOne({_id: new ObjectId(id)})
-        return post? PostMapper(post) : null
+    //Возвращает пост переработанный в мапере
+    static async getPostById(id: string): Promise<OutputPostType | null> {
+        try {
+            if (!isValidObjectId(id)) {
+                throw new Error('id no objectID!');
+            }
+            const post: WithId<PostType> | null = await postCollection.findOne({_id: new ObjectId(id)});
+            return post ? PostMapper(post) : null
+        } catch (error) {
+            console.log(error);
+            return null
+        }
     }
 
-    static async addPost(params: PostCreateModel) : Promise<string> {
-        const blog: OutputBlogType | null = await BlogRepository.getBlogById(params.blogId);
-        const newPost : PostType = {
-            title: params.title,
-            shortDescription: params.shortDescription,
-            content: params.content,
-            blogId: params.blogId,
-            blogName: blog!.name,
-            createdAt: new Date().toISOString()
-        }
-
-        const addResult = await postCollection.insertOne(newPost)
+    // Возвращает ID созданного поста
+    static async addPost(newPost: PostType): Promise<string> {
+        const addResult = await postCollection.insertOne(newPost);
         return addResult.insertedId.toString()
     }
 
-    static async updatePost(params: PostUpdateModel, id: string): Promise<boolean> {
-        const updateResult = await postCollection.updateOne({_id: new ObjectId(id)}, {
-            $set : {
-                title: params.title,
-                shortDescription: params.shortDescription,
-                content: params.content,
-                blogId: params.blogId
+    //Возвращает ✅true (пост найден), ❌false (пост не найден)
+    static async updatePost(updateParams: PostUpdateModel, id: string): Promise<boolean> {
+        try {
+
+            if (!isValidObjectId(id)) {
+                throw new Error("id no objectID!");
             }
-        })
-        return !!updateResult.matchedCount
+
+            const updateResult = await postCollection.updateOne({_id: new ObjectId(id)}, {
+                $set: updateParams
+            });
+            return !!updateResult.matchedCount
+
+        } catch (error) {
+            console.log(error);
+            return false
+        }
+
     }
 
+    // Возарщает ✅true (пост найден и удален), ❌false (пост не найден)
     static async deletePostById(id: string): Promise<boolean> {
-        const deleteResult = await postCollection.deleteOne({_id : new ObjectId(id)})
-        return !!deleteResult.deletedCount
+        try {
+            if (!isValidObjectId(id)) {
+                throw new Error("id no objectID!");
+            }
+            const deleteResult = await postCollection.deleteOne({_id: new ObjectId(id)});
+            return !!deleteResult.deletedCount
+        } catch (error) {
+            console.log(error);
+            return false
+        }
     }
 
+    // ⚠️Удаление всех постов для тестов
     static async deleteAll() {
         await postCollection.deleteMany({})
     }
-
 }
